+++ src/models/certificate-reason.ts
@@ -1,6 +1,6 @@
 import { ReleaseReason } from "./release-reason";
 
 export interface CertificateReason {
-  reasons: ReleaseReason[];
+  releaseReasons: ReleaseReason[];
   releaseDate: Date;
 }
